+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Changelog
 
-## dev-master
+## 1.0.0
+
+ - No changes
+
+## 1.0.0-RC7
 
  - BUGFIX      #189    Fix travis mysql service
+ - FEATURE     #187    Added support of plain text messages
+ - FEATURE     #157    Add reference store and tag cache invalidation
 
 ## 1.0.0-RC6